refactor: move readTemplate to util for use beyond bootstrap
