+++ packages/cli/src/__test__/bootstrap/generators/mergePackageJson.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { readTemplate } from "../../../bootstrap/env/readTemplate";
+import { readTemplate } from "../../../util/readTemplate";
 import { mergePackageJson } from "../../../bootstrap/generators/mergePackageJson";
 import { patchPackageJson } from "../../../bootstrap/generators/patchPackageJson";
